Contribute buffered mode of operation for out_vmware_loginsight
This:
- updates the plugin to use the process/write methods that are available
  since fluentd 1.0 (instead of the 0.12's emit)
- implements the write method so buffered mode is available for the
  plugin
- makes sure that the old sync mode will be used if no buffer
  configuration is specified by the user
- provides some defaults for buffer configs

With buffered mode available, one can make use of the built-in retry
backoff functionality when a request to vRLI fails. Furthermore, it
allows for seamlessly batching requests regardless of whether events
were emitted one by one or in bunches. It also allows to offload the
queue of events to the hard drive, to specify size limits for the buffer
and an overflow policy.

Signed-off-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>